+++ crates/systems/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "namada_systems"
+description = "Namada systems abstract interfaces"
+resolver = "2"
+authors.workspace = true
+edition.workspace = true
+documentation.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+namada_core = { path = "../core" }
+namada_storage = { path = "../storage" }
