+++ crates/systems/src/ethereum_bridge.rs
@@ -1,0 +1,1 @@
+//! Ethereum bridge abstract interfaces
