+++ crates/systems/src/lib.rs
@@ -1,0 +1,27 @@
+//! Abstract interfaces for interacting with Namada systems.
+
+#![doc(html_favicon_url = "https://dev.namada.net/master/favicon.png")]
+#![doc(html_logo_url = "https://dev.namada.net/master/rustdoc-logo.png")]
+#![deny(rustdoc::broken_intra_doc_links)]
+#![deny(rustdoc::private_intra_doc_links)]
+#![warn(
+    missing_docs,
+    rust_2018_idioms,
+    clippy::cast_sign_loss,
+    clippy::cast_possible_truncation,
+    clippy::cast_possible_wrap,
+    clippy::cast_lossless,
+    clippy::arithmetic_side_effects,
+    clippy::dbg_macro,
+    clippy::print_stdout,
+    clippy::print_stderr
+)]
+
+pub mod ethereum_bridge;
+pub mod governance;
+pub mod ibc;
+pub mod parameters;
+pub mod pgf;
+pub mod proof_of_stake;
+pub mod shielded_token;
+pub mod trans_token;
