+++ crates/systems/src/shielded_token.rs
@@ -1,0 +1,1 @@
+//! Shielded token abstract interfaces
